Luizalabs/md-python-dev/requirements.txt
starlette
uvicorn
mongoengine
python-jose
six
dnspython
pytest